+++ libspider/runtime/common/RTFifo.h
@@ -37,7 +37,6 @@
 
 /* === Include(s) === */
 
-
 #include <common/Types.h>
 
 namespace spider {
@@ -54,7 +53,7 @@ namespace spider {
     /* === Class definition === */
 
     struct RTFifo {
-        u64 virtualAddress_ = UINT64_MAX;                   /* = Virtual address of the Fifo = */
+        size_t virtualAddress_ = SIZE_MAX;             /* = Virtual address of the Fifo = */
         u32 size_ = 0;                                      /* = Size of the Fifo = */
         u32 offset_ = 0;                                    /* = Offset in the address = */
         u32 count_ = 1;                                     /* = Number of use of this FIFO = */
